backend/requirements.txt
attrs==26.1.0
bcrypt==5.0.0
blinker==1.9.0
click==8.3.3
colorama==0.4.6
flasgger==0.9.7.1
Flask==3.1.3
flask-cors==6.0.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
mistune==3.2.0
packaging==26.2
PyJWT==2.12.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
six==1.17.0
Werkzeug==3.1.8

